Makefiles finished

+++ src/kernel/systick.h
@@ -1,2 +1,2 @@
-void init_systick(int);
+void init_systick(float);
 void systick_handler(void);
